+++ src/android/ContactsXFindOptions.java
@@ -3,14 +3,16 @@ package de.einfachhans.ContactsX;
 import org.json.JSONObject;
 
 public class ContactsXFindOptions {
-
+    
     boolean displayName = true;
     boolean firstName = true;
     boolean middleName = true;
     boolean familyName = true;
+    boolean organizationName = true;
     boolean phoneNumbers;
     boolean emails;
     boolean addresses;
+    String baseCountryCode = null;
 
     public ContactsXFindOptions(JSONObject options) {
         if (options != null) {
@@ -18,6 +20,12 @@ public class ContactsXFindOptions {
 
             if(fields != null) {
                 this.parseFields(fields);
+            }
+
+            String baseCountryCode = options.optString("baseCountryCode", null);
+
+            if(baseCountryCode != null){
+                this.baseCountryCode = baseCountryCode;
             }
         }
     }
@@ -27,6 +35,7 @@ public class ContactsXFindOptions {
         this.firstName = fields.optBoolean("firstName", true);
         this.middleName = fields.optBoolean("middleName", true);
         this.familyName = fields.optBoolean("familyName", true);
+        this.organizationName = fields.optBoolean("organizationName", true);
         this.phoneNumbers = fields.optBoolean("phoneNumbers");
         this.emails = fields.optBoolean("emails");
         this.addresses = fields.optBoolean("addresses");