+++ src/ios/ContactsXOptions.swift
@@ -3,16 +3,24 @@ class ContactsXOptions {
     var firstName: Bool = true;
     var middleName: Bool = true;
     var familyName: Bool = true;
+    var organizationName: Bool = true;
     var phoneNumbers: Bool = false;
     var emails: Bool = false;
     var addresses: Bool = false;
+    var baseCountryCode : String?? = nil;
 
     init(options: NSDictionary?) {
         if(options != nil) {
-            let fields = options?.value(forKey: "fields") as? NSDictionary ?? nil;
 
+            let fields = options?.value(forKey: "fields") as? NSDictionary ?? nil;
             if(fields != nil) {
                 self.parseFields(fields: fields!)
+            }
+
+            let baseCountryCode = options?.value(forKey: "baseCountryCode") as? String ?? nil;
+            
+            if(baseCountryCode != nil) {
+                self.baseCountryCode = baseCountryCode;
             }
         }
     }
@@ -21,6 +29,7 @@ class ContactsXOptions {
         firstName = fields.value(forKey: "firstName") as? Bool ?? true;
         middleName = fields.value(forKey: "middleName") as? Bool ?? true;
         familyName = fields.value(forKey: "familyName") as? Bool ?? true;
+        organizationName = fields.value(forKey: "organizationName") as? Bool ?? true;
         phoneNumbers = fields.value(forKey: "phoneNumbers") as? Bool ?? false;
         emails = fields.value(forKey: "emails") as? Bool ?? false;
         addresses = fields.value(forKey: "addresses") as? Bool ?? false;
@@ -33,6 +42,7 @@ class ContactXOptions {
     var firstName: String? = nil;
     var middleName: String? = nil;
     var familyName: String? = nil;
+    var organizationName: String? = nil;
     var phoneNumbers: [ContactXValueTypeOptions]? = nil;
     var emails: [ContactXValueTypeOptions]? = nil;
     var addresses: [ContactXAddressOptions]? = nil
@@ -43,6 +53,7 @@ class ContactXOptions {
             firstName = options?.value(forKey: "firstName") as? String;
             middleName = options?.value(forKey: "middleName") as? String;
             familyName = options?.value(forKey: "familyName") as? String;
+            organizationName = options?.value(forKey: "organizationName") as? String;
             let phonenumberArray = options?.value(forKey: "phoneNumbers") as? [NSDictionary];
             if(phonenumberArray != nil) {
                 phoneNumbers = self.parsePhoneNumbers(array: phonenumberArray!);